+++ webhook-bridge/Cargo.toml
@@ -1,8 +1,8 @@
 [workspace]
 members = [
     "generic-queue",
-    "svix-agent-types",
-    "svix-agent",
-    "svix-agent-plugin-generic",
-    "svix-agent-plugin-webhook-receiver",
+    "svix-webhook-bridge-types",
+    "svix-webhook-bridge",
+    "svix-webhook-bridge-plugin-queue-consumer",
+    "svix-webhook-bridge-plugin-webhook-receiver",
 ]
