+++ webhook-bridge/svix-webhook-bridge-plugin-queue-consumer/tests/redis_stream_consumer.rs
@@ -6,11 +6,11 @@ use std::time::Duration;
 use redis::{AsyncCommands, Client};
 use serde_json::json;
 use svix::api::MessageIn;
-use svix_agent_plugin_generic::{
+use svix_webhook_bridge_plugin_queue_consumer::{
     config::{OutputOpts, SvixOptions},
     CreateMessageRequest, RedisConsumerConfig, RedisConsumerPlugin, RedisInputOpts,
 };
-use svix_agent_types::Plugin;
+use svix_webhook_bridge_types::Plugin;
 use wiremock::matchers::method;
 use wiremock::{Mock, MockServer, ResponseTemplate};
