+++ webhook-bridge/svix-webhook-bridge-plugin-webhook-receiver/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "svix-agent-plugin-webhook-receiver"
+name = "svix-webhook-bridge-plugin-webhook-receiver"
 version = "0.1.0"
 edition = "2021"
 
@@ -15,7 +15,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 smol = "1.3"
 svix = "0.84.1"
-svix-agent-types = { path = "../svix-agent-types" }
+svix-webhook-bridge-types = { path = "../svix-webhook-bridge-types" }
 threadpool = "1"
 tracing = "0.1"
 tokio = { version = "1", features = ["full"] }
